add user and logging

+++ express-bugs-server/server.js
@@ -1,15 +1,16 @@
 
-
+import path from 'path'
 import express from 'express'
 import cookieParser from 'cookie-parser'
 import fs from 'fs'
 import PDFDocument from 'pdfkit'
 
 import { bugService } from './services/bug.service.js'
+import { userService } from './services/user.service.js'
 import { loggerService } from './services/logger.service.js'
 
 const app = express()
-const port = 3030
+const PORT = 3030
 
 
 app.use(express.static('public'))
@@ -91,6 +92,9 @@ app.get('/api/bug/:id', (req, res) => {//get bug
 })
 
 app.delete('/api/bug/:id', (req, res) => { //delete bug
+    const loggedinUser = userService.validateToken(req.cookies.loginToken)
+    if (!loggedinUser) return res.status(401).send('Cannot remove bug')
+
     const { bugId } = req.params
 
     bugService.remove(bugId)
@@ -102,7 +106,10 @@ app.delete('/api/bug/:id', (req, res) => { //delete bug
 })
 
 
-app.put('/api/bug', (req, res) => {
+app.put('/api/bug', (req, res) => { //update bug
+    const loggedinUser = userService.validateToken(req.cookies.loginToken)
+    if (!loggedinUser) return res.status(401).send('Cannot update bug')
+
     const { _id, title, description, severity, createdAt, labels } = req.body
     const bugToSave = {
         _id,
@@ -112,7 +119,6 @@ app.put('/api/bug', (req, res) => {
         createdAt: +createdAt || 0,
         labels: labels || []
     }
-    // console.log('bugToSave-server:', bugToSave)
     bugService.save(bugToSave)
         .then(savedBug => res.send(savedBug))
         .catch(err => {
@@ -123,18 +129,24 @@ app.put('/api/bug', (req, res) => {
 })
 
 
-app.post('/api/bug', (req, res) => {
+app.post('/api/bug', (req, res) => { //create bug
+    const loggedinUser = userService.validateToken(req.cookies.loginToken)
+    if (!loggedinUser) return res.status(401).send('Cannot add bug')
     const { title, description, severity, createdAt, labels } = req.body
     const bugToSave = {
         title: title || '',
         description: description || '',
         severity: +severity || 0,
         createdAt: +createdAt || 0,
-        labels: labels || []
+        labels: labels || [],
+        creator: {
+            _id: loggedinUser._id,
+            fullname: loggedinUser.fullname
+        }
     }
 
     // console.log('bugToSave-server:', bugToSave)
-    bugService.save(bugToSave)
+    bugService.save(bugToSave, loggedinUser)
         .then(savedBug => res.send(savedBug))
         .catch(err => {
             loggerService.error(`Couldn't add bug`, err)
@@ -142,4 +154,67 @@ app.post('/api/bug', (req, res) => {
         })
 })
 
-app.listen(port, () => loggerService.info(`Server listening on port http://127.0.0.1:${port}/`))
+// AUTH API
+app.get('/api/user', (req, res) => { //get users
+    userService.query()
+        .then((users) => {
+            res.send(users)
+        })
+        .catch((err) => {
+            console.log('Cannot load users', err)
+            res.status(400).send('Cannot load users')
+        })
+})
+
+app.get('/api/user/:userId', (req, res) => { //get user
+    const { userId } = req.params
+    userService.getById(userId)
+        .then((user) => {
+            res.send(user)
+        })
+        .catch((err) => {
+            console.log('Cannot load user', err)
+            res.status(400).send('Cannot load user')
+        })
+})
+
+app.post('/api/auth/login', (req, res) => {
+    const credentials = req.body
+    userService.checkLogin(credentials)
+        .then(user => {
+            if (user) {
+                const loginToken = userService.getLoginToken(user)
+                res.cookie('loginToken', loginToken)
+                res.send(user)
+            } else {
+                res.status(401).send('Invalid Credentials')
+            }
+        })
+})
+
+app.post('/api/auth/signup', (req, res) => {
+    const credentials = req.body
+    userService.save(credentials)
+        .then(user => {
+            if (user) {
+                const loginToken = userService.getLoginToken(user)
+                res.cookie('loginToken', loginToken)
+                res.send(user)
+            } else {
+                res.status(400).send('Cannot signup')
+            }
+        })
+})
+
+app.post('/api/auth/logout', (req, res) => {
+    res.clearCookie('loginToken')
+    res.send('logged-out!')
+})
+
+app.get('/**', (req, res) => {
+    res.sendFile(path.resolve('public/index.html'))
+})
+
+
+
+app.listen(PORT, () => loggerService.info(`Server listening on port http://127.0.0.1:${PORT}/`))
